Fixes #24
