--- src/gpu_primitives/mod.rs
@@ -1,19 +1,0 @@
-pub trait GpuPrimitive {
-    fn data(&self) -> Vec<u8>;
-    fn desc<'a>() -> wgpu::VertexBufferLayout<'a>;
-}
-
-mod vertex;
-pub use vertex::GpuVertex;
-
-mod quad;
-pub use quad::GpuQuad;
-
-mod camera;
-pub use camera::CameraUniform;
-
-mod transform;
-pub use transform::GpuTransform;
-
-mod world_radius;
-pub use world_radius::WorldUniform;
