+++ src/gpu_types/world.rs
@@ -1,18 +1,42 @@
 use wgpu::{util::DeviceExt, BindGroup, BindGroupLayout, Buffer, Device};
 
-use crate::gpu_primitives::WorldUniform;
+use crate::{
+    gpu_types::GpuUniform,
+    sim::{WORLD_EDGE_SEGMENTS, WORLD_RADIUS},
+};
 
-pub const WORLD_RADIUS: f32 = 50.0;
-pub const WORLD_EDGE_SEGMENTS: u32 = 500;
+#[repr(C)]
+#[derive(Copy, Clone, Debug)]
+pub struct WorldUniform {
+    pub radius: f32,
+    pub boundary_segments: u32,
+    pub padding: [f32; 2],
+}
 
-pub fn bind_world_uniform(
-    device: &Device,
-) -> (Buffer, Vec<u8>, BindGroup, BindGroupLayout) {
-    let layout = create_wradius_bind_group_layout(device);
-    let buffer_contents = get_wradius_buffer_contents();
-    let buffer = create_wradius_buffer(device, &buffer_contents);
-    let bind_group = create_world_bind_group(&buffer, &layout, device);
-    (buffer, buffer_contents, bind_group, layout)
+impl Default for WorldUniform {
+    fn default() -> Self {
+        Self {
+            radius: WORLD_RADIUS,
+            boundary_segments: WORLD_EDGE_SEGMENTS,
+            padding: [f32::default(), f32::default()],
+        }
+    }
+}
+
+unsafe impl bytemuck::Pod for WorldUniform {}
+unsafe impl bytemuck::Zeroable for WorldUniform {}
+
+impl GpuUniform for WorldUniform {
+    fn bind(
+        &self,
+        device: &Device,
+    ) -> (Buffer, Vec<u8>, BindGroup, BindGroupLayout) {
+        let layout = create_wradius_bind_group_layout(device);
+        let buffer_contents = get_wradius_buffer_contents();
+        let buffer = create_wradius_buffer(device, &buffer_contents);
+        let bind_group = create_world_bind_group(&buffer, &layout, device);
+        (buffer, buffer_contents, bind_group, layout)
+    }
 }
 
 fn get_wradius_buffer_contents() -> Vec<u8> {
