+++ src/sim/mod.rs
@@ -1,3 +1,6 @@
+pub const WORLD_RADIUS: f32 = 50.0;
+pub const WORLD_EDGE_SEGMENTS: u32 = 500;
+
 mod body;
 pub use body::Body;
 
@@ -5,5 +8,3 @@ mod state;
 pub use state::State;
 
 pub mod input;
-
-pub mod world;
